requirements.txt
fastapi
uvicorn
transformers
torch
pillow
numpy